[70] - Remove initialization logic of the entry outside of the component
